requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
rich==13.7.1
typer==0.12.3
